+++ src/viewer/VTKWriter.cpp
@@ -14,9 +14,9 @@ VTKWriter::VTKWriter
 
 //============================================================================
 void VTKWriter::renderFrame (
-		double** U,
-		double** V,
-		double** P,
+        REAL** U,
+        REAL** V,
+        REAL** P,
 		int it
 	)
 {
